+++ dpta-service/src/main/java/com/cqjtu/dpta/service/RefundRServiceImpl.java
@@ -2,19 +2,18 @@ package com.cqjtu.dpta.service;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.cqjtu.dpta.api.PafCommService;
+import com.cqjtu.dpta.api.RefundRService;
 import com.cqjtu.dpta.common.util.PageQueryUtil;
 import com.cqjtu.dpta.common.util.PageResult;
-import com.cqjtu.dpta.dao.entity.CommR;
 import com.cqjtu.dpta.dao.entity.PafComm;
 import com.cqjtu.dpta.dao.entity.RefundR;
 import com.cqjtu.dpta.dao.mapper.RefundRMapper;
-import com.cqjtu.dpta.api.RefundRService;
-import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import jakarta.annotation.Resource;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
 import java.util.List;
 
 /**
@@ -50,7 +49,7 @@ public class RefundRServiceImpl extends ServiceImpl<RefundRMapper, RefundR> impl
         for (Long id : ids) {
             QueryWrapper<PafComm> wrapper = new QueryWrapper<>();
             wrapper.eq("refund_id",id);
-            int count = pafCommService.count(wrapper);
+            long count = pafCommService.count(wrapper);
             if (count == 0) {
                 sus += baseMapper.deleteById(id);
             }
